updated gdc

+++ src/main/java/Factorial.java
@@ -1,5 +1,5 @@
 public class Factorial {
-    public static int factorial(int n){
+    public static long factorial(int n){
         if (n == 0) return 1;
         return n * factorial(n - 1);
     }
